Extract handling invoice actions into hook
